app/services/gemini_image_service.py
"""
Gemini 3.0 Pro 이미지 생성 서비스
"""
import os
import base64
from io import BytesIO
from PIL import Image
from google import genai
from app.core.config import GEMINI_API_KEY


async def generate_image_gemini3(prompt: str, base_image: bytes = None) -> str:
    """
    Gemini 3.0 Pro를 사용한 이미지 생성
    
    Args:
        prompt: 이미지 생성 프롬프트 (영어)
        base_image: 기본 이미지 바이트 (image-to-image용, 선택적)
    
    Returns:
        base64 인코딩된 이미지 문자열
    """
    if not GEMINI_API_KEY:
        raise ValueError("GEMINI_API_KEY not configured in .env")
    
    try:
        client = genai.Client(api_key=GEMINI_API_KEY)
        
        # Gemini 3.0 Pro는 이미지 생성 API를 사용
        # 참고: Gemini 3.0 Pro의 실제 이미지 생성 API는 아직 베타일 수 있으므로
        # 현재는 텍스트 기반 프롬프트로 이미지 생성 요청
        
        if base_image:
            # Image-to-Image: 기본 이미지와 프롬프트를 함께 사용
            # PIL Image로 변환
            base_img = Image.open(BytesIO(base_image))
            
            # Gemini 3.0 Pro는 멀티모달 입력을 지원하므로 이미지와 텍스트를 함께 전달
            # 실제 구현은 Gemini API의 최신 문서를 참조해야 함
            # 여기서는 텍스트 기반 생성으로 대체 (실제 API가 준비되면 수정 필요)
            
            # 임시: 이미지가 있으면 프롬프트에 추가 정보 포함
            enhanced_prompt = f"Modify this image based on: {prompt}. Keep the original composition and style while applying the requested changes."
            
            # Gemini 3.0 Pro 이미지 생성 (실제 API 호출)
            # 참고: Gemini 3.0 Pro의 이미지 생성 API는 generate_content를 사용할 수 있음
            # 하지만 현재는 텍스트 생성만 지원하므로, 이미지 생성은 다른 방식으로 처리해야 할 수 있음
            
            # 실제 구현 예시 (API가 준비되면):
            # response = client.models.generate_content(
            #     model="gemini-3.0-pro",
            #     contents=[
            #         {"role": "user", "parts": [
            #             {"text": enhanced_prompt},
            #             {"inline_data": {"mime_type": "image/png", "data": base64.b64encode(base_image).decode()}}
            #         ]}
            #     ]
            # )
            
            # 현재는 FLUX와 유사한 방식으로 처리 (실제 Gemini 3.0 Pro 이미지 생성 API가 준비되면 수정)
            raise NotImplementedError("Gemini 3.0 Pro image-to-image generation is not yet fully implemented. Please use FLUX model for image-to-image.")
        else:
            # Text-to-Image: 프롬프트만 사용
            # Gemini 3.0 Pro의 이미지 생성 기능은 아직 베타 단계이므로
            # 실제 구현은 Google의 최신 문서를 참조해야 함
            
            # 현재는 텍스트 기반 생성으로 처리
            # 실제 API가 준비되면 다음과 같이 구현:
            # response = client.models.generate_content(
            #     model="gemini-3.0-pro",
            #     contents=[{"role": "user", "parts": [{"text": prompt}]}],
            #     config={"response_mime_type": "image/png"}  # 이미지 생성 모드
            # )
            
            # 임시: 에러 메시지 반환 (실제 구현 대기)
            raise NotImplementedError(
                "Gemini 3.0 Pro image generation API is not yet available. "
                "Please use FLUX or SDXL models for image generation. "
                "When Gemini 3.0 Pro image generation becomes available, this will be updated."
            )
        
    except NotImplementedError:
        raise
    except Exception as e:
        raise Exception(f"Gemini 3.0 Pro 이미지 생성 실패: {e}")


async def modify_image_gemini3(base_image: bytes, modification_prompt: str) -> str:
    """
    Gemini 3.0 Pro를 사용한 이미지 수정
    
    Args:
        base_image: 수정할 기본 이미지 바이트
        modification_prompt: 수정 요청 프롬프트 (영어)
    
    Returns:
        base64 인코딩된 이미지 문자열
    """
    # Image-to-Image와 동일한 로직 사용
    return await generate_image_gemini3(modification_prompt, base_image)

